remove helper.js in room-interface
