+++ src/room-interface/helper-RoomInterface.js
@@ -125,3 +125,22 @@ export function getShouldDisabledRedouble(bidSeq) {
   }
   return true;
 }
+
+export function declarerToString(declarer) {
+  switch (declarer) {
+    case 'WEST':
+      return 'W';
+    case 'NORTH':
+      return 'N';
+    case 'EAST':
+      return 'E';
+    case 'SOUTH':
+      return 'S';
+    case 'NS':
+      return 'NS';
+    case 'EW':
+      return 'EW';
+    default:
+      return '';
+  }
+}
